requirements.txt
requests>=2.25.0
pymongo>=4.0.0
python-dotenv>=0.19.0
coverage>=7.0.0 